Change master bathroom fan to 20 minutes before turning off
The bathroom light is often on so using it as the trigger for the
fan isn't reliable. Intsead, change it to a flat "if it's been on
for 20 minutes" as it should be enough time. We can adjust the time
if needed but without a motion sensor in this room using a simple
timer is probably as good as it gets.