Remove automation to turn on outside light in mornings
There's already an automation to turn it on at sunset and another
to turn it off at sunrise. It's not entirely clear what this
automation that is being removed is for.

There's a period of time every spring and fall where the outside
lights are on all day long *because* of this automation. I thought
it was due to a bug.